refactor(dashboard): own Töölaud in app/dashboard with a framework-free service layer (#860)
Relocate the cross-cutting aggregator app/templates/dashboard.py (1,541 lines)
into a proper feature module, app/dashboard/, with a clean layering split so
the data layer is wrappable as Phase-5 REST/MCP tools without the web stack:

  - app/dashboard/service.py  — framework-free data layer. Every raw
    _connect()/SPARQL widget query + bookmark CRUD as inputs->rows functions.
    No fasthtml/starlette imports. Query logic kept byte-equivalent
    (relocation + layering, not a behaviour change).
  - app/dashboard/pages.py    — FastHTML rendering + route handlers +
    register_dashboard_routes, consuming the service functions. The
    ThreadPoolExecutor soft-timeout for the EU-deadlines SPARQL call stays in
    this layer (it calls the synchronous service function).
  - app/dashboard/__init__.py — exports register_dashboard_routes *lazily*
    (PEP 562 __getattr__) so `import app.dashboard.service` does not drag the
    FastHTML page layer into sys.modules.

Service-function inventory (one per widget, all in service.py):
  _get_active_drafter_sessions      — "jätka koostajas"
  _get_high_risk_reports            — "kõrge riskiga leiud"
  _get_unviewed_reports             — "mõjuaruanne valmis, vaatamata"
  _get_stale_analysis_drafts        — "aegunud analüüsid"
  _get_recent_syncs                 — "uued ontoloogia muudatused"
  _get_recent_exports               — "hiljutised ekspordid"
  _get_awaiting_review_drafts       — reviewer "ülevaatuse järgi ootavad" (#817)
  _get_unresolved_annotation_drafts — "eelnõud lahtiste märkustega"
  _get_eu_transposition_deadlines   — A6 EU deadlines (soft-timeout wrapper)
  _get_bookmarks / _get_user_org_info
  _add_bookmark / _remove_bookmark  — bookmark CRUD

Bookmark add/remove routes move into pages.py with their #743 XHR/JSON contract
intact (X-Requested-With -> JSON 200/401, plain form -> 303) and #848 URI
scheme validation; pinned by the existing route tests.

app/main.py repointed to `from app.dashboard import register_dashboard_routes`;
registration order unchanged. app/templates/dashboard.py + .gitkeep deleted and
the now-empty app/templates/ directory removed. No re-export shim — zero
functional templates.dashboard references remain in app/ or tests/.

Patch-path updates (patch-where-used):
  - widget loaders + dashboard_page + capability helpers + bookmark routes
    + log_action -> app.dashboard.pages.*
  - _add_bookmark/_remove_bookmark unit tests + _connect -> app.dashboard.service.*
  - _get_eu_transposition_deadlines + list_overdue_or_upcoming_transpositions
    -> app.dashboard.service.*
  - test_import_safety PAGE_MODULES: app.templates.dashboard -> app.dashboard.pages

Adds tests/test_dashboard_import_direction.py: AST scan asserting service.py's
source imports no fasthtml/starlette (mirroring test_ui_import_direction's
style), plus a subprocess guard that the package __init__ stays lazy.

Verification: ruff check + pyright clean repo-wide; full suite 5012 passed,
50 skipped (no regression vs ~4920 baseline).

Part of #860.
